abstract color platte.
